+++ src/domain/entity/entity.ts
@@ -1,4 +1,4 @@
-import { UniqueEntityIdVO } from "./common/value-object";
+import { UniqueEntityIdVO } from "../common/value-object";
 
 export abstract class Entity {
   readonly id: UniqueEntityIdVO;
