+++ src/domain/entity/index.ts
@@ -1,0 +1,2 @@
+export * from './assignor.entity';
+export * from './receivable.entity';
